refactor: move it tests to test tree
