+++ webapi/BUILD.bazel
@@ -188,61 +188,3 @@ scala_library(
         "@maven//:redis_clients_jedis",
     ],
 )
-
-# Should be removed, after each scala_test rule is refactored to use direct
-# dependencies to the needed packages.
-scala_library(
-    name = "it_library",
-    testonly = True,
-    srcs = [
-        "//webapi/src/it/scala/org/knora/webapi:src",
-    ],
-    resources = [
-        "//sipi/config",
-        "//webapi/src/it/resources",
-    ],
-    unused_dependency_checker_mode = "warn",
-    runtime_deps = [
-        "@maven//:ch_qos_logback_logback_classic",
-        "@maven//:ch_qos_logback_logback_core",
-        "@maven//:com_typesafe_akka_akka_slf4j_2_12",
-        "@maven//:org_slf4j_log4j_over_slf4j",
-    ],
-    deps = [
-        "//webapi:test_library",
-        "//webapi/src/main/scala/org/knora/webapi",
-        "//webapi/src/main/scala/org/knora/webapi/app",
-        "//webapi/src/main/scala/org/knora/webapi/core",
-        "//webapi/src/main/scala/org/knora/webapi/exceptions",
-        "//webapi/src/main/scala/org/knora/webapi/feature",
-        "//webapi/src/main/scala/org/knora/webapi/instrumentation",
-        "//webapi/src/main/scala/org/knora/webapi/messages",
-        "//webapi/src/main/scala/org/knora/webapi/routing",
-        "//webapi/src/main/scala/org/knora/webapi/settings",
-        "//webapi/src/main/scala/org/knora/webapi/util",
-        # Logging
-        "@maven//:com_typesafe_scala_logging_scala_logging_2_12",
-        "@maven//:org_slf4j_slf4j_api",
-        # Akka
-        "@maven//:com_typesafe_akka_akka_actor_2_12",
-        "@maven//:com_typesafe_akka_akka_http_2_12",
-        "@maven//:com_typesafe_akka_akka_http_core_2_12",
-        "@maven//:com_typesafe_akka_akka_http_spray_json_2_12",
-        "@maven//:com_typesafe_akka_akka_parsing_2_12",
-        "@maven//:com_typesafe_akka_akka_stream_2_12",
-        "@maven//:com_typesafe_config",
-        "@maven//:io_spray_spray_json_2_12",
-        "@maven//:org_scala_lang_modules_scala_xml_2_12",
-        "@maven//:org_scala_lang_scala_library",
-        "@maven//:org_scala_lang_scala_reflect",
-        "@maven//:org_scalactic_scalactic_2_12",
-        "@maven//:org_scalatest_scalatest_core_2_12",
-        "@maven//:org_scalatest_scalatest_wordspec_2_12",
-        "@maven//:org_scalatest_scalatest_matchers_core_2_12",
-        "@maven//:org_scalatest_scalatest_shouldmatchers_2_12",
-        "@maven//:org_scalatest_scalatest_compatible",
-        "@maven//:org_testcontainers_testcontainers",
-        "@maven//:junit_junit",
-        "@maven//:org_xmlunit_xmlunit_core",
-    ],
-)
