--- webapi/src/it/resources/BUILD.bazel
@@ -1,12 +1,0 @@
-package(default_visibility = ["//visibility:public"])
-
-filegroup(
-    name = "resources",
-    srcs = [
-        "fuseki.conf",
-        "graphdb-free.conf",
-        "graphdb-se.conf",
-        "logback-test.xml",
-        "test.conf",
-    ],
-)
