--- webapi/src/it/resources/fuseki.conf
@@ -1,16 +1,0 @@
-include "test"
-
-app {
-    triplestore {
-        dbtype = "fuseki"
-
-        fuseki {
-            port = 3030
-            repository-name = "knora-test-unit"
-            username = "admin"
-            password = "test"
-        }
-
-        reload-on-start = false
-    }
-}
