--- webapi/src/it/resources/graphdb-se.conf
@@ -1,16 +1,0 @@
-include "test"
-
-app {
-    triplestore {
-        dbtype = "graphdb-se"
-
-        graphdb {
-            port = 7200
-            repository-name = "knora-test-unit"
-            username = "admin"
-            password = "root"
-        }
-
-        reload-on-start = false
-    }
-}
