--- webapi/src/it/resources/logback-test.xml
@@ -1,96 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!--
-  ~ Copyright © 2015-2019 the contributors (see Contributors.md).
-  ~
-  ~ This file is part of Knora.
-  ~
-  ~ Knora is free software: you can redistribute it and/or modify
-  ~ it under the terms of the GNU Affero General Public License as published
-  ~ by the Free Software Foundation, either version 3 of the License, or
-  ~ (at your option) any later version.
-  ~
-  ~ Knora is distributed in the hope that it will be useful,
-  ~ but WITHOUT ANY WARRANTY; without even the implied warranty of
-  ~ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-  ~ GNU Affero General Public License for more details.
-  ~
-  ~ You should have received a copy of the GNU Affero General Public
-  ~ License along with Knora.  If not, see <http://www.gnu.org/licenses/>.
-  -->
-
-<!-- For assistance related to logback-translator or configuration  -->
-<!-- files in general, please contact the logback user mailing list -->
-<!-- at http://www.qos.ch/mailman/listinfo/logback-user             -->
-<!--                                                                -->
-<!-- For professional support please see                            -->
-<!--    http://www.qos.ch/shop/products/professionalSupport         -->
-<!--                                                                -->
-<configuration scan="false" debug="false">
-    <appender name="stdout" class="ch.qos.logback.core.ConsoleAppender">
-        <encoder>
-            <pattern>[%date{ISO8601}] %-5level - %logger{0} - %msg%n</pattern>
-        </encoder>
-    </appender>
-<!--    <appender name="R" class="ch.qos.logback.core.rolling.RollingFileAppender">-->
-<!--        <File>logs/console.log</File>-->
-<!--        <encoder>-->
-<!--            <pattern>[%date{ISO8601}] %-5level %logger{0} - %X{sourceThread} - %X{akkaSource} - %msg%n</pattern>-->
-<!--        </encoder>-->
-<!--        <rollingPolicy class="ch.qos.logback.core.rolling.TimeBasedRollingPolicy">-->
-<!--            <fileNamePattern>logs/console.log.%d{yyyy-MM-dd-HH}</fileNamePattern>-->
-<!--        </rollingPolicy>-->
-<!--    </appender>-->
-
-    <logger name="org.apache.jena.riot" level="ERROR"/>
-    <logger name="org.apache.jena.tdb" level="ERROR"/>
-    <logger name="akka" level="INFO"/>
-    <logger name="akka.event.slf4j.Slf4jLogger" level="ERROR"/>
-    <logger name="net.sf.ehcache.config" level="ERROR"/>
-    <logger name="kamon.metric" level="ERROR"/>
-
-    <!-- V1 -->
-
-    <logger name="org.knora" level="INFO"/>
-    <logger name="org.knora.webapi" level="INFO"/>
-    <logger name="org.knora.webapi.util.cache" level="INFO"/>
-    <logger name="org.knora.webapi.util.PermissionUtilADM" level="INFO"/>
-    <logger name="org.knora.webapi.update.UpdateRepository" level="INFO"/>
-    <logger name="org.knora.webapi.routing.Authenticator" level="INFO"/>
-    <logger name="org.knora.webapi.routing.v1.AssetsRouteV1" level="INFO"/>
-    <logger name="org.knora.webapi.routing.v1.ResourcesRouteV1" level="INFO"/>
-    <logger name="org.knora.webapi.routing.v1.StandoffRouteV1" level="INFO"/>
-    <logger name="org.knora.webapi.routing.v1.ValuesRouteV1" level="INFO"/>
-    <logger name="org.knora.webapi.store.StoreManager" level="INFO"/>
-    <logger name="org.knora.webapi.store.triplestore" level="INFO"/>
-    <logger name="org.knora.webapi.store.triplestore.http.HttpTriplestoreConnector" level="INFO"/>
-    <logger name="org.knora.webapi.responders.v1.CkanResponderV1" level="INFO"/>
-    <logger name="org.knora.webapi.responders.v1.UsersResponderV1" level="INFO"/>
-    <logger name="org.knora.webapi.responders.v1.ProjectsResponderV1" level="INFO"/>
-    <logger name="org.knora.webapi.responders.v1.PermissionsResponderV1" level="INFO"/>
-    <logger name="org.knora.webapi.responders.v1.ResourcesResponderV1" level="INFO"/>
-    <logger name="org.knora.webapi.responders.v1.ValuesResponderV1" level="INFO"/>
-    <logger name="org.knora.webapi.responders.v1.StoreResponderV1" level="INFO"/>
-    <logger name="org.knora.webapi.responders.v1.OntologyResponderV1" level="INFO"/>
-    <logger name="org.knora.webapi.viewhandlers.ResourceHtmlView" level="INFO"/>
-    <logger name="org.knora.webapi.other.v1.DrawingsGodsV1ITSpec" level="INFO"/>
-    <logger name="org.knora.webapi.e2e.v1.KnoraSipiIntegrationV1ITSpec" level="INFO"/>
-    <logger name="org.knora.webapi.e2e.v1.KnoraSipiScriptsV1ITSpec" level="INFO"/>
-
-    <!-- Spec Logging -->
-    <logger name="org.knora.webapi.ITKnoraLiveSpec" level="INFO"/>
-    <logger name="org.knora.webapi.ITKnoraFakeSpec" level="INFO"/>
-
-    <!-- metrics logging (info level turns them on, error level turns them off -->
-    <logger name="M-Main$" level="ERROR"/>
-    <logger name="M-UsersResponderADM" level="ERROR"/>
-    <logger name="M-HttpTriplestoreConnector" level="ERROR"/>
-    <logger name="M-CacheServiceManager" level="ERROR"/>
-    <logger name="M-CacheSerialization" level="ERROR"/>
-
-
-    <root level="INFO">
-        <appender-ref ref="stdout"/>
-<!--        <appender-ref ref="R"/>-->
-    </root>
-</configuration>
