--- webapi/src/it/resources/test.conf
@@ -1,4 +1,0 @@
-include "application"
-
-app.print-short-config = false
-app.triplestore.auto-init = true
