+++ webapi/src/test/scala/org/knora/webapi/it/v1/BUILD.bazel
@@ -3,16 +3,11 @@ package(default_visibility = ["//visibility:public"])
 load("@io_bazel_rules_scala//scala:scala.bzl", "scala_test")
 load("//third_party:dependencies.bzl", "ALL_WEBAPI_MAIN_DEPENDENCIES", "BASE_TEST_DEPENDENCIES", "BASE_TEST_DEPENDENCIES_WITH_JSON", "BASE_TEST_DEPENDENCIES_WITH_JSON_LD")
 
-filegroup(
-    name = "src",
-    srcs = glob(["**/*.scala"]),
-)
-
 scala_test(
-    name = "VersionRouteITSpec",
+    name = "DrawingsGodsV1ITSpec",
     size = "small",  # 60s
     srcs = [
-        "e2e/VersionRouteITSpec.scala",
+        "DrawingsGodsV1ITSpec.scala",
     ],
     data = [
         "//knora-ontologies",
@@ -21,7 +16,7 @@ scala_test(
     jvm_flags = ["-Dconfig.resource=fuseki.conf"],
     # unused_dependency_checker_mode = "warn",
     deps = ALL_WEBAPI_MAIN_DEPENDENCIES + [
-        "//webapi:it_library",
+        "//webapi:test_library",
         "//webapi:main_library",
     ] + BASE_TEST_DEPENDENCIES_WITH_JSON,
 )
@@ -30,7 +25,7 @@ scala_test(
     name = "ErrorV1ITSpec",
     size = "small",  # 60s
     srcs = [
-        "e2e/v1/ErrorV1ITSpec.scala",
+        "ErrorV1ITSpec.scala",
     ],
     data = [
         "//knora-ontologies",
@@ -39,7 +34,7 @@ scala_test(
     jvm_flags = ["-Dconfig.resource=fuseki.conf"],
     # unused_dependency_checker_mode = "warn",
     deps = ALL_WEBAPI_MAIN_DEPENDENCIES + [
-        "//webapi:it_library",
+        "//webapi:test_library",
         "//webapi:main_library",
     ] + BASE_TEST_DEPENDENCIES_WITH_JSON,
 )
@@ -48,7 +43,7 @@ scala_test(
     name = "KnoraSipiIntegrationV1ITSpec",
     size = "medium",
     srcs = [
-        "e2e/v1/KnoraSipiIntegrationV1ITSpec.scala",
+        "KnoraSipiIntegrationV1ITSpec.scala",
     ],
     data = [
         "//knora-ontologies",
@@ -57,9 +52,8 @@ scala_test(
     jvm_flags = ["-Dconfig.resource=fuseki.conf"],
     # unused_dependency_checker_mode = "warn",
     deps = ALL_WEBAPI_MAIN_DEPENDENCIES + [
-        "//webapi:it_library",
-        "//webapi:main_library",
         "//webapi:test_library",
+        "//webapi:main_library",
         "@maven//:org_scala_lang_modules_scala_xml_2_12",
         "@maven//:org_xmlunit_xmlunit_core",
     ] + BASE_TEST_DEPENDENCIES_WITH_JSON,
@@ -69,7 +63,7 @@ scala_test(
     name = "KnoraSipiPermissionsV1ITSpec",
     size = "small",  # 60s
     srcs = [
-        "e2e/v1/KnoraSipiPermissionsV1ITSpec.scala",
+        "KnoraSipiPermissionsV1ITSpec.scala",
     ],
     data = [
         "//knora-ontologies",
@@ -78,7 +72,7 @@ scala_test(
     jvm_flags = ["-Dconfig.resource=fuseki.conf"],
     # unused_dependency_checker_mode = "warn",
     deps = ALL_WEBAPI_MAIN_DEPENDENCIES + [
-        "//webapi:it_library",
+        "//webapi:test_library",
         "//webapi:main_library",
     ] + BASE_TEST_DEPENDENCIES_WITH_JSON,
 )
@@ -87,7 +81,7 @@ scala_test(
     name = "KnoraSipiScriptsV1ITSpec",
     size = "small",  # 60s
     srcs = [
-        "e2e/v1/KnoraSipiScriptsV1ITSpec.scala",
+        "KnoraSipiScriptsV1ITSpec.scala",
     ],
     data = [
         "//knora-ontologies",
@@ -96,44 +90,7 @@ scala_test(
     jvm_flags = ["-Dconfig.resource=fuseki.conf"],
     # unused_dependency_checker_mode = "warn",
     deps = ALL_WEBAPI_MAIN_DEPENDENCIES + [
-        "//webapi:it_library",
-        "//webapi:main_library",
-    ] + BASE_TEST_DEPENDENCIES_WITH_JSON,
-)
-
-scala_test(
-    name = "KnoraSipiIntegrationV2ITSpec",
-    size = "small",  # 60s
-    srcs = [
-        "e2e/v2/KnoraSipiIntegrationV2ITSpec.scala",
-    ],
-    data = [
-        "//knora-ontologies",
-        "//test_data",
-    ],
-    jvm_flags = ["-Dconfig.resource=fuseki.conf"],
-    # unused_dependency_checker_mode = "warn",
-    deps = ALL_WEBAPI_MAIN_DEPENDENCIES + [
-        "//webapi:it_library",
-        "//webapi:main_library",
         "//webapi:test_library",
-    ] + BASE_TEST_DEPENDENCIES_WITH_JSON,
-)
-
-scala_test(
-    name = "DrawingsGodsV1ITSpec",
-    size = "small",  # 60s
-    srcs = [
-        "other/v1/DrawingsGodsV1ITSpec.scala",
-    ],
-    data = [
-        "//knora-ontologies",
-        "//test_data",
-    ],
-    jvm_flags = ["-Dconfig.resource=fuseki.conf"],
-    # unused_dependency_checker_mode = "warn",
-    deps = ALL_WEBAPI_MAIN_DEPENDENCIES + [
-        "//webapi:it_library",
         "//webapi:main_library",
     ] + BASE_TEST_DEPENDENCIES_WITH_JSON,
 )
